Fix what clang reports that gcc does not
gcc and clang do not warn about the same things, and building the tree with
clang and -Werror turns up 66 errors gcc is silent about. One of them is a
defect:

    moolineview.cpp:512: error: variable 'space' is used uninitialized
      whenever 'if' condition is false [-Wsometimes-uninitialized]

check_if_scrolled() reads the parent's scrollbar-spacing into a local. The
GTK+2 branch asks for it unconditionally, since a widget with a visible
horizontal scrollbar has a parent. The GTK+3 branch, marked "written by AI",
added a NULL check around the call but left the read of the value outside it,
so a NULL parent puts whatever was on the stack into the delta calculation.
gtk_widget_get_parent() exists in GTK+2.24 as well, so both branches collapse
into one that initialises the variable and keeps the check; the FIXME marker
goes with the #if.

The other 65 are not defects, but they are what stands between the tree and a
clang build:

MOO_DEFINE_SLIST, _DLIST and _QUEUE in moolist.h generate a family of static
inline helpers -- 59 of the errors -- of which each user calls two or three.
gcc does not warn about an unused static inline; clang does, once the macro is
expanded in a .c file. They are G_GNUC_UNUSED now, with the continuation
backslashes still at column 73.

Five upper-bound guards compare an enum against the count of its own values,
which in C++ is provably true because the type carries the range: type <
N_TOOLS in moousertools.cpp and le < G_N_ELEMENTS (line_end_menu_items) in
mooeditwindow.cpp. Casting to an integer keeps the guard live for a value cast
into the enum from outside, which is what it is there for.

get_folds_in_range() in moofold.cpp has no caller but itself. clang reports it
as "not needed and will not be emitted", which is exactly what it is; deleting
it changes no binary. Removing it exposed two more, both of which had to stay:
get_line_count() is called only from g_assert(), which a release build compiles
out, and is marked G_GNUC_UNUSED; moo_cairo_create() has all three of its
callers inside #if !GTK_CHECK_VERSION(3,0,0), so its definition moves under the
same guard rather than being marked.

Verified: clang 19 with -DENABLE_STRICT=ON now finishes at zero errors and zero
non-deprecation warnings, and gcc still builds both toolkits clean.

+++ src/mooedit/mooeditwindow.cpp
@@ -1454,7 +1454,7 @@ update_doc_line_end_item (MooEditWindow *window)
     g_return_if_fail (action != nullptr);
 
     le = moo_edit_get_line_end_type (doc);
-    g_return_if_fail (le > 0 && le < G_N_ELEMENTS(line_end_menu_items));
+    g_return_if_fail (le > 0 && (guint) le < G_N_ELEMENTS (line_end_menu_items));
 
     moo_menu_mgr_set_active (moo_menu_action_get_mgr (MOO_MENU_ACTION (action)),
                              line_end_menu_items[le], TRUE);
